scheduler: fixed initial CR feedback. Still need to refactor
