+++ controller/scheduler-v2/formation.go
@@ -52,17 +52,9 @@ func (f *Formation) Update(procs map[string]int) map[string]int {
 	return diff
 }
 
-type formationCounts map[utils.FormationKey]map[string]int
+type formationJobs map[utils.FormationKey]map[string]int
 
-func NewFormationCounts(fs Formations) formationCounts {
-	fc := make(formationCounts)
-	for fKey := range fs {
-		fc[fKey] = make(map[string]int)
-	}
-	return fc
-}
-
-func (fc formationCounts) AddJob(j *Job) {
+func (fc formationJobs) AddJob(j *Job) {
 	key := j.Formation.key()
 	f, ok := fc[key]
 	if !ok {
